New LabBridge/labbridge/requirements.txt
reflex>=0.8.0
pdfplumber>=0.11.0
pandas>=2.1.0
google-genai>=0.2.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-multipart>=0.0.9
cloudinary>=1.36.0
supabase>=2.0.0
python-dotenv>=1.0.0
Unidecode>=1.3.6
tenacity>=8.2.0
httpx>=0.25.0
fastapi>=0.110.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-mock>=3.11.0
playwright>=1.40.0
stripe>=7.0.0
resend>=0.7.0
